+++ openscm/errors.py
@@ -58,4 +58,3 @@ class OverwriteError(AssertionError):
     """
 
     pass
-