+++ tests/unit/test_highlevel.py
@@ -368,24 +368,28 @@ def test_filter_by_regexp(test_scm_df):
 
 def test_filter_timeseries_different_length():
     df = ScmDataFrame(
-        pd.DataFrame(np.array([[1.0, 2.0, 3.0], [4.0, 5.0, np.nan]]).T, index=[2000, 2001, 2002]),
+        pd.DataFrame(
+            np.array([[1.0, 2.0, 3.0], [4.0, 5.0, np.nan]]).T, index=[2000, 2001, 2002]
+        ),
         columns={
-            'model': ['a_iam'],
-            'climate_model': ['a_model'],
-            'scenario': ['a_scenario', 'a_scenario2'],
-            'region': ['World'],
-            'variable': ['Primary Energy'],
-            'unit': ['EJ/y']
-        }
+            "model": ["a_iam"],
+            "climate_model": ["a_model"],
+            "scenario": ["a_scenario", "a_scenario2"],
+            "region": ["World"],
+            "variable": ["Primary Energy"],
+            "unit": ["EJ/y"],
+        },
     )
 
-    npt.assert_array_equal(df.filter(scenario='a_scenario2').timeseries().squeeze(), [4.0, 5.0])
+    npt.assert_array_equal(
+        df.filter(scenario="a_scenario2").timeseries().squeeze(), [4.0, 5.0]
+    )
     npt.assert_array_equal(df.filter(year=2002).timeseries().squeeze(), 3.0)
 
-    exp = pd.Series(['a_scenario'], name='scenario')
-    obs = df.filter(year=2002)['scenario']
+    exp = pd.Series(["a_scenario"], name="scenario")
+    obs = df.filter(year=2002)["scenario"]
     pd.testing.assert_series_equal(exp, obs)
-    assert df.filter(scenario='a_scenario2', year=2002).timeseries().empty
+    assert df.filter(scenario="a_scenario2", year=2002).timeseries().empty
 
 
 def test_timeseries(test_scm_df):